moved AddProjects container to ListProjects folder
