+++ app/containers/AccountSettings/actions.js
@@ -14,7 +14,6 @@ import {
   LISTED_USER_SKILLS_ACTION,
   DELETE_USER_SKILLS_ACTION,
   DELETED_USER_SKILLS_ACTION
-
 } from './constants';
 
 export function defaultAction() {
