+++ app/containers/ListCommunications/index.js
@@ -4,7 +4,7 @@
  *
  */
 
-import React from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Helmet } from 'react-helmet';
@@ -20,7 +20,7 @@ import reducer from './reducer';
 import saga from './saga';
 import { listCommunication } from './actions';
 
-export class ListCommunications extends React.Component { // eslint-disable-line react/prefer-stateless-function
+export class ListCommunications extends Component { // eslint-disable-line react/prefer-stateless-function
   constructor(props){
     super(props)
   }
