+++ app/containers/ListProjects/index.js
@@ -12,7 +12,7 @@ import { createStructuredSelector } from 'reselect';
 import { compose } from 'redux';
 
 import ContentWrapper from 'components/Layout/ContentWrapper';
-import { Row, Col, Button, Table, Pagination,Modal } from 'react-bootstrap';
+import { Row, Col, Button, Table, Pagination, Modal } from 'react-bootstrap';
 
 import injectSaga from 'utils/injectSaga';
 import injectReducer from 'utils/injectReducer';
@@ -22,8 +22,8 @@ import saga from './saga';
 
 import ReactQuill from 'react-quill';
 import 'react-quill/dist/quill.snow.css';
-import * as a from "./actions";
-import AddProject from 'containers/AddProject';
+import * as a from './actions';
+import AddProject from './AddProject';
 import { styles, headings } from 'assets/styles/variables';
